src/main/java/org/openlmis/referencedata/service/ProgramOrderableService.java
/*
 * This program is part of the OpenLMIS logistics management information system platform software.
 * Copyright © 2017 VillageReach
 *
 * This program is free software: you can redistribute it and/or modify it under the terms
 * of the GNU Affero General Public License as published by the Free Software Foundation, either
 * version 3 of the License, or (at your option) any later version.
 *  
 * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
 * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
 * See the GNU Affero General Public License for more details. You should have received a copy of
 * the GNU Affero General Public License along with this program. If not, see
 * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
 */

package org.openlmis.referencedata.service;

import org.openlmis.referencedata.domain.Program;
import org.openlmis.referencedata.domain.ProgramOrderable;
import org.openlmis.referencedata.repository.ProgramOrderableRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProgramOrderableService {

  @Autowired
  private ProgramOrderableRepository programOrderableRepository;

  /**
   * Finds ProgramOrderables matching all of provided parameters.
   * @param program program of searched ProgramOrderables.
   * @return list of all ProgramOrderables matching all of provided parameters.
   */
  public List<ProgramOrderable> searchProgramOrderables(Program program) {
    return programOrderableRepository.searchProgramOrderables(program);
  }
}
